remove all legacy gsheets and gapi code
