+++ src/background/background.ts
@@ -32,7 +32,6 @@ import initContextMenus from "@/background/contextMenus";
 import initBrowserAction from "@/background/browserAction";
 import initInstaller from "@/background/installer";
 import initNavigation from "@/background/navigation";
-import initGoogle, { isGAPISupported } from "@/contrib/google/initGoogle";
 import initExecutor from "@/background/executor";
 import initBrowserCommands from "@/background/initBrowserCommands";
 import initDeploymentUpdater from "@/background/deploymentUpdater";
@@ -66,13 +65,3 @@ initStarterMods();
 initPartnerTokenRefresh();
 initLogSweep();
 initModUpdater();
-
-if (isGAPISupported()) {
-  // Optimistically initialize Google API, if Google API is supported. But do not prompt for permissions
-  void initGoogle();
-} else {
-  console.debug("Google API not supported by browser", {
-    // @ts-expect-error -- exists on Chromium, but not other browsers
-    browserBrands: navigator.userAgentData?.brands,
-  });
-}
