+++ src/pageEditor/pageEditor.tsx
@@ -29,7 +29,6 @@ import ReactDOM from "react-dom";
 import React from "react";
 import Panel from "@/pageEditor/Panel";
 import { watchNavigation } from "@/pageEditor/protocol";
-import initGoogle from "@/contrib/google/initGoogle";
 import { initToaster } from "@/utils/notify";
 import { markAppStart } from "@/utils/performance";
 import { initRuntimeLogging } from "@/development/runtimeLogging";
@@ -39,7 +38,6 @@ markAppStart();
 void initMessengerLogging();
 void initRuntimeLogging();
 registerMessenger();
-void initGoogle();
 watchNavigation();
 initToaster();
