+++ src/sidebar/sidebar.tsx
@@ -31,7 +31,6 @@ import React from "react";
 import registerBuiltinBlocks from "@/bricks/registerBuiltinBlocks";
 import registerContribBlocks from "@/contrib/registerContribBlocks";
 import { initToaster } from "@/utils/notify";
-import initGoogle from "@/contrib/google/initGoogle";
 import { initRuntimeLogging } from "@/development/runtimeLogging";
 
 function init(): void {
@@ -41,7 +40,6 @@ function init(): void {
 void initMessengerLogging();
 void initRuntimeLogging();
 registerMessenger();
-void initGoogle();
 registerContribBlocks();
 registerBuiltinBlocks();
 initToaster();
